exercises part 4

+++ controllers/blogs.js
@@ -1,27 +1,32 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const middleware = require('../utils/middleware')
 
-blogsRouter.get('/', (request, response) => {
-  Blog.find({}).then((blogs) => {
-    response.json(blogs)
-  })
+blogsRouter.get('/', async (request, response) => {
+  const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
+  response.json(blogs)
 })
 
-blogsRouter.post('/', (request, response) => {
+blogsRouter.post('/', middleware.userExtractor, async (request, response) => {
   if (!request.body.url || !request.body.title) {
-    response.status(400).end()
-  } else {
-    const blog = new Blog({
-      title: request.body.title,
-      url: request.body.url,
-      author: request.body.author,
-      likes: request.body.likes || 0,
-    })
-
-    blog.save().then((result) => {
-      response.status(201).json(result)
-    })
+    return response.status(400).end()
   }
+  if (!request.token) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  const user = request.user
+  const blog = new Blog({
+    title: request.body.title,
+    url: request.body.url,
+    author: request.body.author,
+    likes: request.body.likes || 0,
+    user: user._id,
+  })
+
+  const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
+  response.status(201).json(savedBlog)
 })
 
 blogsRouter.get('/:id', async (request, response) => {
@@ -33,10 +38,25 @@ blogsRouter.get('/:id', async (request, response) => {
   }
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
-})
+blogsRouter.delete(
+  '/:id',
+  middleware.userExtractor,
+  async (request, response) => {
+    const user = request.user
+    const blog = await Blog.findById(request.params.id)
+    if (blog) {
+      if (blog.user.toString() !== user.id.toString()) {
+        console.log('blog.user.toString()', blog.user.toString())
+        return response.status(401).json({ error: 'token invalid' })
+      } else {
+        await Blog.findByIdAndDelete(request.params.id)
+      }
+      response.status(204).end()
+    } else {
+      response.status(404).end()
+    }
+  }
+)
 
 blogsRouter.put('/:id', (request, response, next) => {
   const body = request.body
